Add support for range sensor to turn on the display
Integrate the HC-SR04 range sensor into the mirror.  First, make sure
the screensaver is enabled so that the screen will blank after a while
(to save energy when no on is around).

Then, when the some object is close enough, turn on the display.

While adding this, we also changed the frequency at which the clock
display is updated from 200 ms(!) to 500 ms.  The clock only shows
minutes so 200 ms seems like way overkill.  But we want the sensor to
quickly detect objects, so do the measurements every 500 ms.